feat(dsg): controller base
