+++ packages/amplication-data-service-generator/src/server/resource/controller/controller.base.template.ts
@@ -1,0 +1,244 @@
+import * as common from "@nestjs/common";
+import * as swagger from "@nestjs/swagger";
+import * as nestMorgan from "nest-morgan";
+import * as nestAccessControl from "nest-access-control";
+// @ts-ignore
+import * as basicAuthGuard from "../auth/basicAuth.guard";
+// @ts-ignore
+import * as abacUtil from "../auth/abac.util";
+// @ts-ignore
+import { isRecordNotFoundError } from "../prisma.util";
+// @ts-ignore
+import * as errors from "../errors";
+
+declare interface CREATE_QUERY {}
+declare interface UPDATE_QUERY {}
+declare interface DELETE_QUERY {}
+
+declare interface CREATE_INPUT {}
+declare interface WHERE_INPUT {}
+declare interface WHERE_UNIQUE_INPUT {}
+declare interface UPDATE_INPUT {}
+
+declare const FINE_ONE_PATH: string;
+declare const UPDATE_PATH: string;
+declare const DELETE_PATH: string;
+declare interface FIND_ONE_QUERY {}
+
+declare class ENTITY {}
+declare interface Select {}
+
+declare interface SERVICE {
+  create(args: { data: CREATE_INPUT; select: Select }): Promise<ENTITY>;
+  findMany(args: { where: WHERE_INPUT; select: Select }): Promise<ENTITY[]>;
+  findOne(args: {
+    where: WHERE_UNIQUE_INPUT;
+    select: Select;
+  }): Promise<ENTITY | null>;
+  update(args: {
+    where: WHERE_UNIQUE_INPUT;
+    data: UPDATE_INPUT;
+    select: Select;
+  }): Promise<ENTITY>;
+  delete(args: { where: WHERE_UNIQUE_INPUT; select: Select }): Promise<ENTITY>;
+}
+
+declare const RESOURCE: string;
+declare const ENTITY_NAME: string;
+declare const CREATE_DATA_MAPPING: CREATE_INPUT;
+declare const UPDATE_DATA_MAPPING: UPDATE_INPUT;
+declare const SELECT: Select;
+
+export class CONTROLLER_BASE {
+  constructor(
+    protected readonly service: SERVICE,
+    protected readonly rolesBuilder: nestAccessControl.RolesBuilder
+  ) {}
+
+  @common.UseInterceptors(nestMorgan.MorganInterceptor("combined"))
+  @common.UseGuards(basicAuthGuard.BasicAuthGuard, nestAccessControl.ACGuard)
+  @common.Post()
+  @nestAccessControl.UseRoles({
+    resource: ENTITY_NAME,
+    action: "create",
+    possession: "any",
+  })
+  @swagger.ApiCreatedResponse({ type: ENTITY })
+  @swagger.ApiForbiddenResponse({ type: errors.ForbiddenException })
+  async create(
+    @common.Query() query: CREATE_QUERY,
+    @common.Body() data: CREATE_INPUT,
+    @nestAccessControl.UserRoles() userRoles: string[]
+  ): Promise<ENTITY> {
+    const permission = this.rolesBuilder.permission({
+      role: userRoles,
+      action: "create",
+      possession: "any",
+      resource: ENTITY_NAME,
+    });
+    const invalidAttributes = abacUtil.getInvalidAttributes(permission, data);
+    if (invalidAttributes.length) {
+      const properties = invalidAttributes
+        .map((attribute: string) => JSON.stringify(attribute))
+        .join(", ");
+      const roles = userRoles
+        .map((role: string) => JSON.stringify(role))
+        .join(",");
+      throw new errors.ForbiddenException(
+        `providing the properties: ${properties} on ${ENTITY_NAME} creation is forbidden for roles: ${roles}`
+      );
+    }
+    // @ts-ignore
+    return await this.service.create({
+      ...query,
+      data: CREATE_DATA_MAPPING,
+      select: SELECT,
+    });
+  }
+
+  @common.UseInterceptors(nestMorgan.MorganInterceptor("combined"))
+  @common.UseGuards(basicAuthGuard.BasicAuthGuard, nestAccessControl.ACGuard)
+  @common.Get()
+  @nestAccessControl.UseRoles({
+    resource: ENTITY_NAME,
+    action: "read",
+    possession: "any",
+  })
+  @swagger.ApiOkResponse({ type: [ENTITY] })
+  @swagger.ApiForbiddenResponse()
+  async findMany(
+    @common.Query() query: WHERE_INPUT,
+    @nestAccessControl.UserRoles() userRoles: string[]
+  ): Promise<ENTITY[]> {
+    const permission = this.rolesBuilder.permission({
+      role: userRoles,
+      action: "read",
+      possession: "any",
+      resource: ENTITY_NAME,
+    });
+    const results = await this.service.findMany({
+      where: query,
+      select: SELECT,
+    });
+    return results.map((result) => permission.filter(result));
+  }
+
+  @common.UseInterceptors(nestMorgan.MorganInterceptor("combined"))
+  @common.UseGuards(basicAuthGuard.BasicAuthGuard, nestAccessControl.ACGuard)
+  @common.Get(FINE_ONE_PATH)
+  @nestAccessControl.UseRoles({
+    resource: ENTITY_NAME,
+    action: "read",
+    possession: "own",
+  })
+  @swagger.ApiOkResponse({ type: ENTITY })
+  @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
+  @swagger.ApiForbiddenResponse({ type: errors.ForbiddenException })
+  async findOne(
+    @common.Query() query: FIND_ONE_QUERY,
+    @common.Param() params: WHERE_UNIQUE_INPUT,
+    @nestAccessControl.UserRoles() userRoles: string[]
+  ): Promise<ENTITY | null> {
+    const permission = this.rolesBuilder.permission({
+      role: userRoles,
+      action: "read",
+      possession: "own",
+      resource: ENTITY_NAME,
+    });
+    const result = await this.service.findOne({
+      ...query,
+      where: params,
+      select: SELECT,
+    });
+    if (result === null) {
+      throw new errors.NotFoundException(
+        `No resource was found for ${JSON.stringify(params)}`
+      );
+    }
+    return permission.filter(result);
+  }
+
+  @common.UseInterceptors(nestMorgan.MorganInterceptor("combined"))
+  @common.UseGuards(basicAuthGuard.BasicAuthGuard, nestAccessControl.ACGuard)
+  @common.Patch(UPDATE_PATH)
+  @nestAccessControl.UseRoles({
+    resource: ENTITY_NAME,
+    action: "update",
+    possession: "any",
+  })
+  @swagger.ApiOkResponse({ type: ENTITY })
+  @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
+  @swagger.ApiForbiddenResponse({ type: errors.ForbiddenException })
+  async update(
+    @common.Query() query: UPDATE_QUERY,
+    @common.Param() params: WHERE_UNIQUE_INPUT,
+    @common.Body()
+    data: UPDATE_INPUT,
+    @nestAccessControl.UserRoles() userRoles: string[]
+  ): Promise<ENTITY | null> {
+    const permission = this.rolesBuilder.permission({
+      role: userRoles,
+      action: "update",
+      possession: "any",
+      resource: ENTITY_NAME,
+    });
+    const invalidAttributes = abacUtil.getInvalidAttributes(permission, data);
+    if (invalidAttributes.length) {
+      const properties = invalidAttributes
+        .map((attribute: string) => JSON.stringify(attribute))
+        .join(", ");
+      const roles = userRoles
+        .map((role: string) => JSON.stringify(role))
+        .join(",");
+      throw new errors.ForbiddenException(
+        `providing the properties: ${properties} on ${ENTITY_NAME} update is forbidden for roles: ${roles}`
+      );
+    }
+    try {
+      return await this.service.update({
+        ...query,
+        where: params,
+        data: UPDATE_DATA_MAPPING,
+        select: SELECT,
+      });
+    } catch (error) {
+      if (isRecordNotFoundError(error)) {
+        throw new errors.NotFoundException(
+          `No resource was found for ${JSON.stringify(params)}`
+        );
+      }
+      throw error;
+    }
+  }
+
+  @common.UseInterceptors(nestMorgan.MorganInterceptor("combined"))
+  @common.UseGuards(basicAuthGuard.BasicAuthGuard, nestAccessControl.ACGuard)
+  @common.Delete(DELETE_PATH)
+  @nestAccessControl.UseRoles({
+    resource: ENTITY_NAME,
+    action: "delete",
+    possession: "any",
+  })
+  @swagger.ApiOkResponse({ type: ENTITY })
+  @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
+  @swagger.ApiForbiddenResponse({ type: errors.ForbiddenException })
+  async delete(
+    @common.Query() query: DELETE_QUERY,
+    @common.Param() params: WHERE_UNIQUE_INPUT
+  ): Promise<ENTITY | null> {
+    try {
+      return await this.service.delete({
+        ...query,
+        where: params,
+        select: SELECT,
+      });
+    } catch (error) {
+      if (isRecordNotFoundError(error)) {
+        throw new errors.NotFoundException(
+          `No resource was found for ${JSON.stringify(params)}`
+        );
+      }
+      throw error;
+    }
+  }
+}
